restructure cli
