+++ packages/cli/src/commands/canonicalSwap/index.ts
@@ -1,0 +1,6 @@
+export * from './getCanonicalBalance';
+export * from './getWormholeBalance';
+export * from './swapCanWormhole';
+export * from './swapWormholeCan';
+export * from './swapV2toV3';
+
