+++ packages/cli/src/commands/index.ts
@@ -1,0 +1,3 @@
+export * from './token';
+export * from './canonicalSwap';
+export * from './tbc'
