+++ packages/cli/src/commands/token/index.ts
@@ -1,0 +1,5 @@
+export * from './createToken';
+export * from './addMetadata';
+export * from './getTokenInfo';
+export * from './freezeAccount';
+
