+++ packages/cli/src/utils/config.ts
@@ -1,0 +1,12 @@
+import * as anchor from "@project-serum/anchor";
+
+export const config = {
+    cswap: {
+        canonicalMintV3: 'sRLY3migNrkC1HLgqotpvi66qGkdNedqPZ9TJpAQhyh',
+        canonicalDataV3: '97exq8nMj13ydaQAwxMN26nfHXtvoVioexYkoB1ZUs2v',
+        canonicalMintV2: 'RLYv2ubRMDLcGG2UyvPmnPmkfuQTsMbg4Jtygc7dmnq',
+        canonicalDataV2: '61SVmuBgYEChu3vBR4PXtTzHbXYsn78CM7uZCSquieHu',
+        wormholeMint: '6Y7LNYkHiJHSH8zR2HvZQzXD3QA9yFw64tyMHxBxDRe4',
+        wormholeData: 'FvdfyPydxRgCnFPwBdrx7B1fuBAZFxWCahATumYvzEdv'
+    }
+};
