+++ packages/ts/package.json
@@ -37,6 +37,7 @@
     "@types/chai": "^4.3.0",
     "@types/mocha": "^9.0.0",
     "chai": "^4.3.4",
+    "chai-as-promised": "^7.1.1",
     "mocha": "^8.4.0",
     "ts-mocha": "^8.0.0",
     "typescript": "^4.5.4"
